:bug: Rename module to revocation_setup.py
